Add class for action and announce the main functions
Class for action is the 'class action'. It involves attribute that is verb, subject and object of action, etc., and functions for to get some attribute.

Function get_actions will get list of action found in tree that was built on sentence. Other function is help function for 'get_actions' and they will be written after.