Add some information for action and change dependences classification
Add sentence text to root of tree constructed on this and to action extracted from this. Also add name for action. Now any action has name 'Action{verb_of_action}'

Change dependences classifiaction: all words that was not classificated as subject, object or not informative words will be classificated as 'OTHER'.